+++ xxcmd/consoleui.py
@@ -1,6 +1,7 @@
 # consoleui.py
 import os
 import curses
+from curses.ascii import isprint
 import time
 from .lineedit import LineEdit
 
@@ -260,11 +261,8 @@ class ConsoleUI():
         # Check for key press event handler
         elif key in self.key_events.keys():
             self.key_events[key]()
-        # Ignore strange stuff
-        elif len(key) > 1:
-            pass
-        # Just record the key press
-        else:
+        # Add printable ascii characters to our input
+        elif len(key) == 1 and isprint(key):
             self.input.addchar(key)
 
         # Trigger other event handlers